Componentize the product form product tags field
